14-08 Actualizacion presentacion editar  listado

+++ app/src/main/java/com/example/almacen/feature_activity/data/remote/dto/ActivityDetailRequest.kt
@@ -1,7 +1,7 @@
 package com.example.almacen.feature_activity.data.remote.dto
 
 data class ActivityDetailRequest(
-    val idArticulo: Long,
+    val idArticulo: Int,
     val nroLote: String,
     val peso: Double,
     val cajas: Int
